Added check if called function exists
There is no simple way of making a pass over the AST and
deciding whether a function has been defined for a specific
function call, because all nodes are really specific and do
not have generic linkage to child nodes. Links are described
in the particular node types themselves.

Alternatively, when generating code for a function call,
all function definitions are known. It hence makes sense
to check the function's existence there. However, that
forces the codegen() function to throw Exceptions, which
in turn needs to be applied to all other implementations
of that function as well.