Added return statement constraints in function body
Since no control structures are implemented yet, the only
place returns are allowed to appear (and must, since each
function must return something) is the end of a function.

If control structures are introduced, each code block must
lead to a return statement somehow.

If a return statement is not the last statement of the function
body or the function does not even have any, an exception is thrown.