Made function declarations detached
During function block generation, the old context is stored,
the function block is constructed and assigned to the environment
and the old context is restored.

Hence, the function block is a separate unit only accessible
using a CALL instruction and the statements following the
function declaration are now correctly assigned to the
surrounding context.